+++ protocol/ObjectIds.go
@@ -1,0 +1,6 @@
+package protocol
+
+// ObjectIds is a simple list of object identifiers
+type ObjectIds struct {
+	ObjectIds []string `json:"objectIds"`
+}